feat(cb2-10672): refactor and move action to more appropriate location
