+++ src/app/store/retry-interceptor/actions/retry-interceptor.actions.ts
@@ -1,0 +1,3 @@
+import { createAction, props } from '@ngrx/store';
+
+export const retryInterceptorFailure = createAction('[retry-interceptor] retryInterceptorFailure', props<{ error: string }>());
