+++ src/app/store/retry-interceptor/retry-interceptor.module.ts
@@ -1,0 +1,8 @@
+import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+
+@NgModule({
+  declarations: [],
+  imports: [CommonModule],
+})
+export class RetryInterceptorStateModule {}
