+++ src/app/store/technical-records/actions/technical-record-service.actions.ts
@@ -106,8 +106,6 @@ export const generateADRCertificate = createAction(`${prefix} generateADRCertifi
 export const generateADRCertificateSuccess = createAction(`${prefix} generateADRCertificate Success`, props<{ id: string }>());
 export const generateADRCertificateFailure = createOutcomeAction('generateADRCertificate', false);
 
-export const retryInterceptorFailure = createAction(`${prefix} retryInterceptorFailure`, props<{ error: string }>());
-
 export const generateContingencyADRCertificate = createAction(`${prefix} generateContingencyADRCertificate`, props<{
   systemNumber: string, createdTimestamp: string, certificateType: string
 }>());
